requirements.txt
Django
drf-yasg
django-extensions
djangorestframework
djangorestframework-simplejwt
